+++ CHANGELOG.md
@@ -37,10 +37,12 @@
     - show
   - snowflake.snowpark.DataFrameWriter:
     - save_as_table
+- Added support for snow:// URLs to `snowflake.snowpark.Session.file.get` and `snowflake.snowpark.Session.file.get_stream`
 
 ### Bug Fixes
 
 - Fixed a bug in local testing that null filled columns for constant functions.
+- Fixed a bug causing `snowflake.snowpark.Session.file.get_stream` to fail for quoted stage locations
 
 ## 1.14.0 (2024-03-20)